==> 00 Add weather headers.ipynb <==
import pandas as pd
import os

obs_dir = 'Data/Weather/raw/Observations/'

out_dir = 'Data/Weather/raw_with_headers/'

if not os.path.exists(out_dir):
    os.makedirs(out_dir)

for filename in os.listdir(obs_dir):
    if not filename.startswith('.'): # ignore hidden files
        f = os.path.join(obs_dir, filename)
        df = pd.read_csv(f)
        df.to_csv(out_dir+'NOAA_daily_weather_observations_'+filename, 
                  header=['ID', 'DATE', 'ELEMENT', 'DATA_VALUE', 'M_FLAG', 'Q_FLAG', 'S_FLAG', 'OBS_TIME'], index=False)

df=pd.read_csv('Data/Weather/raw/ghcnd-stations.csv', header=None, usecols=[0,1,2,3,4,5]) # There are some later columns that we don't need
df.to_csv(out_dir+'NOAA_ghcnd_stations.csv', header=['ID', 'LATITUTDE', 'LONGITUDE', 'ELEVATION', 'STATE', 'NAME'], index=False)

df=pd.read_csv('Data/Weather/raw/ghcnd-inventory.csv', header=None)
df.to_csv(out_dir+'NOAA_ghcnd_inventory.csv', header=['ID', 'LATITUDE', 'LONGITUDE', 'ELEMENT', 'FIRST_YEAR', 'LAST_YEAR'], index=False)
