==> 02 Merge_stations_and_observations.ipynb <==
import pandas as pd
import os
stations=pd.read_csv('Data/Weather/raw_with_headers/NOAA_ghcnd_stations_CA_blanks_removed.csv', usecols=[1,2,3,4,5,6,7])

in_dir = 'Data/Weather/raw_with_headers/'
out_dir = 'Data/Weather/Stations-Observations Merged/'

if not os.path.exists(out_dir):
    os.makedirs(out_dir)

for x in range(2018, 2023):   
    infilename = in_dir + 'NOAA_daily_weather_observations_' + str(x) + '.csv'
    outfilename = out_dir + 'NOAA_daily_weather_observations_' + str(x) + '_with_stations.csv'
    obs=pd.read_csv(infilename)
    station_obs=obs.merge(stations, how='left', on='ID')
    station_obs.dropna(subset=['NAME'], inplace=True) # Drop the rows that are not in California
    station_obs.to_csv(outfilename, index=False)

# create one superfile with all the data
obs18 = pd.read_csv(in_dir + 'NOAA_daily_weather_observations_2018.csv')
obs19 = pd.read_csv(in_dir + 'NOAA_daily_weather_observations_2019.csv')
obs20 = pd.read_csv(in_dir + 'NOAA_daily_weather_observations_2020.csv')
obs21 = pd.read_csv(in_dir + 'NOAA_daily_weather_observations_2021.csv')
obs22 = pd.read_csv(in_dir + 'NOAA_daily_weather_observations_2022.csv')

covidfile = pd.concat([obs20, obs21, obs22])
superfile = pd.concat([obs18, obs19, obs20, obs21, obs22])
superfile.shape

covid_obs=covidfile.merge(stations, how='left', on='ID')
covid_obs.dropna(subset=['NAME'], inplace=True) # Drop the rows that are not in California
covid_obs.to_csv(out_dir + 'NOAA_daily_weather_observations_with_stations_2020_to_2022.csv', index=False)

station_obs=superfile.merge(stations, how='left', on='ID')
station_obs.dropna(subset=['NAME'], inplace=True) # Drop the rows that are not in California
station_obs.to_csv(out_dir + 'NOAA_daily_weather_observations_with_stations_superfile.csv', index=False)
