==> 04 Add FIPS.ipynb <==
import pandas as pd
fips=pd.read_csv('Data/Counties/FIPS_County_Codes.csv', usecols=[0,1])

fips['State']='CA'

covid_in_dir='Data/Covid Cases/'
covid_out_dir='Data/Covid Cases/Covid Cases with FIPS/'
if not os.path.exists(covid_out_dir):
    os.makedirs(covid_out_dir)

vacc_in_dir='Data/COVID-19 Vaccination Data/'
vacc_out_dir='Data/COVID-19 Vaccination Data/COVID-19 Vaccination Data with FIPS/'
if not os.path.exists(vacc_out_dir):
    os.makedirs(vacc_out_dir)

weather_in_dir='Data/Weather/Stations-Observations Merged-Converted/'
weather_out_dir='Data/Weather/Stations-Observations Merged-Converted-with FIPS/'
if not os.path.exists(weather_out_dir):
    os.makedirs(weather_out_dir)

### COVID data

df=pd.read_excel(covid_in_dir+'Covid Cases Reduced2.xlsx', usecols=[1,2,3,4,5,6,7]) # strip first column as it was just an index
df=df.merge(fips, on='County', how='left')
df=df[['Date','County','State','FIPS','Cases','Deaths','TotalTests','PositiveTest']]
df.to_csv(covid_out_dir+'Covid_Cases_Reduced_with_FIPS.csv', index=False)

df

df=pd.read_csv(covid_in_dir+'Statewide COVID-19 Cases Deaths Tests updated.csv')
df.rename(columns={'area':'County'}, inplace=True)
df=df.merge(fips, on='County', how='left')
df=df[['date', 'County', 'State', 'FIPS', 'area_type', 'population', 'cases', 'cumulative_cases', 'deaths', 'cumulative_deaths', 'total_tests',
       'cumulative_total_tests', 'positive_tests', 'cumulative_positive_tests', 'reported_cases', 'cumulative_reported_cases', 'reported_deaths',
       'cumulative_reported_deaths', 'reported_tests']]
df.to_csv(covid_out_dir+'Statewide_COVID-19_Cases_Deaths_Tests_updated_with_FIPS.csv', index=False)

### Vaccination data

df=pd.read_csv(vacc_in_dir+'clean_covid19vaccinesbycounty v2.csv')
df.rename(columns={'county':'County'}, inplace=True)
df=df.merge(fips, on='County', how='left')
df=df[['County', 'State', 'FIPS', 'administered_date', 'total_doses', 'partially_vaccinated',
       'fully_vaccinated', 'at_least_one_dose', 'booster_recip_count']]
df.to_csv(vacc_out_dir+'clean_covid-19_vaccines_by_county_with_FIPS.csv', index=False)

df=pd.read_csv(vacc_in_dir+'covid19vaccinesbycounty.csv')
df.rename(columns={'county':'County'}, inplace=True)
df=df.merge(fips, on='County', how='left')
df=df[['County', 'State', 'FIPS', 'administered_date', 'total_doses', 'cumulative_total_doses',
       'pfizer_doses', 'cumulative_pfizer_doses', 'moderna_doses',
       'cumulative_moderna_doses', 'jj_doses', 'cumulative_jj_doses',
       'partially_vaccinated', 'total_partially_vaccinated',
       'fully_vaccinated', 'cumulative_fully_vaccinated', 'at_least_one_dose',
       'cumulative_at_least_one_dose', 'california_flag',
       'booster_recip_count', 'cumulative_booster_recip_count']]
df.to_csv(vacc_out_dir+'covid-19_vaccines_by_county_with_FIPS.csv', index=False)

### Weather

fips.rename(columns={'County':'COUNTY', 'State':'STATE'}, inplace=True)


import os
for filename in os.listdir(weather_in_dir):
    if not filename.startswith('.'): # ignore hidden files
        f = os.path.join(weather_in_dir, filename)
        if os.path.isfile(f):
            df=pd.read_csv(f)
            df=df.merge(fips, on='COUNTY', how='left')
            file = filename[:-4]
            df.to_csv(weather_out_dir+file+'_with_FIPS.csv', index=False)
