==> 05 Graph exploration.ipynb <==
import pandas as pd
from matplotlib import pyplot as plt

weather = pd.read_csv('Data/Weather/Stations-Observations Merged-Converted-with FIPS/NOAA_daily_weather_observations_with_stations_2020_to_2022_converted_with_FIPS.csv')
weather['DATE']=pd.to_datetime(weather['DATE'], format='%Y%m%d')

max_temps=weather[weather['ELEMENT']=='TMAX_F']
min_temps=weather[weather['ELEMENT']=='TMIN_F']

temps= pd.concat((max_temps, min_temps), ignore_index=True)
avgtemps = pd.DataFrame(temps.groupby(['DATE','COUNTY'])['DATA_VALUE'].mean())
avgtemps = avgtemps.reset_index()  

cases = pd.read_csv('Data/Covid Cases/Covid Cases with FIPS/Covid_Cases_Reduced_with_FIPS.csv')
cases['Date']=pd.to_datetime(cases['Date'], format='%Y-%m-%d')

def simple_graph(county):
    plt.bar(max_temps[max_temps['COUNTY']==county]['DATE'], max_temps[max_temps['COUNTY']==county]['DATA_VALUE'])
    plt.bar(min_temps[min_temps['COUNTY']==county]['DATE'], min_temps[min_temps['COUNTY']==county]['DATA_VALUE'])
    plt.bar(cases[cases['County']==county]['Date'], cases[cases['County']==county]['Cases'])
    #plt.legend(['Max Temp', 'Min Temp'])
    plt.figure(figsize=(2,2))
    plt.show()

def double_bar_hi_lo_temp(county):
 
    # Creating plot with dataset_1
    fig, ax1 = plt.subplots()

    ax1.set_xlabel('date')
    ax1.set_ylabel('temp', color='tab:red')
    ax1.bar(max_temps[max_temps['COUNTY']==county]['DATE'], max_temps[max_temps['COUNTY']==county]['DATA_VALUE'], color = 'tab:red')
    ax1.bar(min_temps[min_temps['COUNTY']==county]['DATE'], min_temps[min_temps['COUNTY']==county]['DATA_VALUE'], color = 'tab:blue')
    plt.xticks(rotation=45, ha='right')
    ax1.tick_params(axis ='y')

    # Adding Twin Axes to plot using dataset_2
    ax2 = ax1.twinx()

    ax2.set_ylabel('cases', color='tab:green')
    ax2.bar(cases[cases['County']==county]['Date'], cases[cases['County']==county]['Cases'], color = 'tab:green')
    ax2.tick_params(axis ='y')


    # Adding title
    plt.title('Use different y-axes on the left and right of a Matplotlib plot', fontweight ="bold")

    # Show plot
    fig.savefig('Graph/'+county+'_bar.png', bbox_inches='tight')
    
    plt.show()

def double_line_avg_temp(county):
    
    # Creating plot with dataset_1
    fig, ax1 = plt.subplots()
    
    ax1.set_xlabel('date')
    ax1.set_ylabel('Avg Temp', color='tab:blue')
    ax1.plot(avgtemps[avgtemps['COUNTY']== county]['DATE'], avgtemps[avgtemps['COUNTY']== county]['DATA_VALUE'], color = 'tab:blue')
    plt.xticks(rotation=45, ha='right')
    ax1.tick_params(axis ='y')
    
    # Adding Twin Axes to plot using dataset_2
    ax2 = ax1.twinx()
    
 
    ax2.set_ylabel('cases', color='tab:green')
    #ax2.set_ylabel('deaths', color='tab:red')
    ax2.plot(cases[cases['County']== county]['Date'], cases[cases['County']== county]['Cases'], color = 'tab:green')
    ax2.plot(cases[cases['County']== county]['Date'], cases[cases['County']== county]['Deaths'], color = 'tab:red')

    ax2.tick_params(axis ='y')


    # Adding title
    plt.title(county, fontweight ="bold")

    # Show plot
    fig.savefig('Graph/Full/'+county+'_line.png', dpi = 300, bbox_inches='tight')
    plt.show()

list_of_counties = avgtemps['COUNTY'].unique()

for county in list_of_counties:
    double_line_avg_temp(county)
