==> 06 Descriptives.ipynb <==
import pandas as pd

weather = pd.read_csv('Data/Weather/Stations-Observations Merged-Converted-with FIPS/NOAA_daily_weather_observations_with_stations_2020_to_2022_converted_with_FIPS.csv')
weather['DATE']=pd.to_datetime(weather['DATE'], format='%Y%m%d')
max_temps=weather[weather['ELEMENT']=='TMAX_F']
min_temps=weather[weather['ELEMENT']=='TMIN_F']
temps= pd.concat((max_temps, min_temps), ignore_index=True)
temps = temps.sort_values(['ID','DATE'], ignore_index=True)


temps

temps['ID'].nunique()

temps['DATE'].max()

temps['DATE'].min()

temps['COUNTY'].nunique()

temps.shape

temps['NAME'].nunique()

temps.to_csv('Data/Weather/Stations-Observations Merged-Converted-with FIPS/Temps for descriptives.csv')

test = temps.loc['2020-01-01':'2020-03-31']

temps.iterrows()

test = temps.set_index(['DATE']).copy()

temps['QUARTER'] = pd.DatetimeIndex(temps['DATE']).quarter
temps['YEAR'] = pd.DatetimeIndex(temps['DATE']).year

temps

Q120 = temps.loc[(temps['YEAR']==2020) & (temps['QUARTER']==1)].reset_index(drop=True)
Q220 = temps.loc[(temps['YEAR']==2020) & (temps['QUARTER']==2)].reset_index(drop=True)
Q320 = temps.loc[(temps['YEAR']==2020) & (temps['QUARTER']==3)].reset_index(drop=True)
Q420 = temps.loc[(temps['YEAR']==2020) & (temps['QUARTER']==4)].reset_index(drop=True)
Q121 = temps.loc[(temps['YEAR']==2021) & (temps['QUARTER']==1)].reset_index(drop=True)
Q221 = temps.loc[(temps['YEAR']==2021) & (temps['QUARTER']==2)].reset_index(drop=True)
Q321 = temps.loc[(temps['YEAR']==2021) & (temps['QUARTER']==3)].reset_index(drop=True)
Q421 = temps.loc[(temps['YEAR']==2021) & (temps['QUARTER']==4)].reset_index(drop=True)
Q122 = temps.loc[(temps['YEAR']==2022) & (temps['QUARTER']==1)].reset_index(drop=True)

Y2020 = temps.loc[(temps['YEAR']==2020)].reset_index(drop=True)
Y2021 = temps.loc[(temps['YEAR']==2021)].reset_index(drop=True)
Y2022 = temps.loc[(temps['YEAR']==2022)].reset_index(drop=True)


Q120.ID.nunique()

list = [Q120, Q220, Q320, Q420, Q121, Q221, Q321, Q421, Q122]
print('Stations')

for item in list:
    print(item.ID.nunique())
print('Counties')

for item in list:
    print(item.COUNTY.nunique())

temps.DATE.max()

q120

Y2020.to_csv('Data/Descriptives/weather2020.csv')
Y2021.to_csv('Data/Descriptives/weather2021.csv')
Y2022.to_csv('Data/Descriptives/weather2022.csv')

cases = pd.read_csv('Data/Covid Cases/Covid Cases with FIPS/Covid_Cases_Reduced_with_FIPS.csv')
cases['Date']=pd.to_datetime(cases['Date'], format='%Y-%m-%d')
cases

cases['QUARTER'] = pd.DatetimeIndex(cases['Date']).quarter
cases['YEAR'] = pd.DatetimeIndex(cases['Date']).year

C2020 = cases.loc[(cases['YEAR']==2020)].reset_index(drop=True)
C2021 = cases.loc[(cases['YEAR']==2021)].reset_index(drop=True)
C2022 = cases.loc[(cases['YEAR']==2022)].reset_index(drop=True)

C2020

C2020.to_csv('Data/Descriptives/cases2020.csv')
C2021.to_csv('Data/Descriptives/cases2021.csv')
C2022.to_csv('Data/Descriptives/cases2022.csv')
